+++ fsleyes/parseargs.py
@@ -464,7 +464,10 @@ OPTIONS = td.TypeDict({
                        'showGridLines',
                        'highlightSlice',
                        'sliceLocation',
-                       'labelSpace'],
+                       'labelSpace',
+                       'reverseSlices',
+                       'sliceOverlap',
+                       'reverseOverlap'],
     'Scene3DOpts'   : ['zoom',
                        'showLegend',
                        'light',
@@ -846,6 +849,9 @@ ARGUMENTS = td.TypeDict({
     'LightBoxOpts.highlightSlice' : ('hs', 'highlightSlice', False),
     'LightBoxOpts.sliceLocation'  : ('ll', 'sliceLocation',  False),
     'LightBoxOpts.labelSpace'     : ('sp', 'labelSpace',     True),
+    'LightBoxOpts.reverseSlices'  : ('rs', 'reverseSlices',  False),
+    'LightBoxOpts.sliceOverlap'   : ('so', 'sliceOverlap',   True),
+    'LightBoxOpts.reverseOverlap' : ('ro', 'reverseOverlap', False),
     'LightBoxOpts.zax'            : ('zx', 'zaxis',          True),
 
     'Scene3DOpts.zoom'           : ('z',   'zoom',           True),
@@ -1128,24 +1134,36 @@ HELP = td.TypeDict({
     'OrthoOpts.ycentre'     : 'Y canvas centre ([-1, 1])',
     'OrthoOpts.zcentre'     : 'Z canvas centre ([-1, 1])',
 
-    'LightBoxOpts.sliceSpacing'   : 'Slice spacing',
-    'LightBoxOpts.numSlices'      : 'Number of slices. Ignored if '
-                                    '--sliceSpacing is specified.',
-    'LightBoxOpts.ncols'          :
+    'LightBoxOpts.sliceSpacing' :
+    'Slice spacing',
+    'LightBoxOpts.numSlices' :
+    'Number of slices. Ignored if --sliceSpacing is specified.',
+    'LightBoxOpts.ncols' :
     'Number of columns. Only used for off-screen rendering.',
-    'LightBoxOpts.nrows'          :
+    'LightBoxOpts.nrows' :
     'Number of rows. Only used for off-screen rendering. If both --ncols and '
     '--nrows are specified, nrows may be adjusted to honour the --zrange and '
     '--sliceSpacing settings.',
-    'LightBoxOpts.zrange'         : 'Slice range',
-    'LightBoxOpts.showGridLines'  : 'Show grid lines',
-    'LightBoxOpts.highlightSlice' : 'Highlight current slice',
-    'LightBoxOpts.sliceLocation'  : 'Show location in world coordinates on '
-                                    'each slice. Ignored if --labelSpace is '
-                                    'specified',
-    'LightBoxOpts.labelSpace'     : 'Show slice locations in this coordinate '
-                                    'system.',
-    'LightBoxOpts.zax'            : 'Z axis',
+    'LightBoxOpts.zrange' :
+    'Slice range',
+    'LightBoxOpts.showGridLines' :
+    'Show grid lines',
+    'LightBoxOpts.highlightSlice' :
+    'Highlight current slice',
+    'LightBoxOpts.sliceLocation' :
+    'Show location in world coordinates on each slice. Ignored if '
+    '--labelSpace is specified',
+    'LightBoxOpts.labelSpace' :
+    'Show slice locations in this coordinate system.',
+    'LightBoxOpts.reverseSlices' :
+    'Display slices from high to low Z value, instead of low to high.',
+    'LightBoxOpts.sliceOverlap' :
+    'Overlap adjacent slices by this much, specified as a percentage.',
+    'LightBoxOpts.reverseOverlap' :
+    'Draw lower slices on top of higher slices instead of higher on top of '
+    'lower.',
+    'LightBoxOpts.zax' :
+    'Z axis',
 
     'Scene3DOpts.zoom'          : 'Zoom (1-5000, default: 100)',
     'Scene3DOpts.showLegend'    : 'Hide the orientation legend',