+++ fsleyes/strings.py
@@ -1036,6 +1036,10 @@ properties = TypeDict({
     'LightBoxOpts.highlightSlice' : 'Highlight slice',
     'LightBoxOpts.showGridLines'  : 'Show grid lines',
     'LightBoxOpts.sliceSpacing'   : 'Slice spacing',
+    'LightBoxOpts.sliceOverlap'   : 'Slice overlap',
+    'LightBoxOpts.reverseSlices'  : 'Reverse slice order',
+    'LightBoxOpts.reverseOverlap' : 'Reverse slice overlap',
+    'LightBoxOpts.sliceOverlap'   : 'Slice overlap',
     'LightBoxOpts.zrange'         : 'Z range',
     'LightBoxOpts.labelSpace'     : 'Show slice location',